Fix format warning in multi_options.cpp

+++ code/network/multi_options.cpp
@@ -206,10 +206,10 @@ void multi_options_read_config()
 							mprintf(("ERROR: Invalid or out of range webapi_server_port '%s' specified in multi.cfg, must be integer between 1024 and %i.\n", tok, USHRT_MAX));
 						}
 						else if(result < 1024) {
-							mprintf(("ERROR: webapi_server_port '%d' in multi.cfg is too low, must be between 1024 and %d.\n", result, static_cast<int>(USHRT_MAX)));
+							mprintf(("ERROR: webapi_server_port '%ld' in multi.cfg is too low, must be between 1024 and %d.\n", result, USHRT_MAX));
 						}
 						else {
-							mprintf(("Using webapi_server_port '%d' from multi.cfg.\n", result));
+							mprintf(("Using webapi_server_port '%ld' from multi.cfg.\n", result));
 							Multi_options_g.webapiPort = (ushort) result;
 						}
 					}
